requirements.txt
alembic==0.8.9
click==6.7            
Flask==0.12           
Flask-Assets==0.12    
Flask-Babel==0.11.1
Flask-Script==2.0.5
httplib2==0.9.2       
itsdangerous==0.24    
Jinja2==2.9.4         
Mako==1.0.6           
MarkupSafe==0.23      
oauth2==1.9.0.post1
pycparser==2.14       
python-editor==1.0.3  
requests==2.12.4
webassets==0.12.0
Werkzeug==0.11.15     
gunicorn==19.6.0
beautifulsoup4==4.5.3
line-bot-sdk==1.1.0